Add script to run katas
Running a kata really means running the unit test for that
kata. This script runs either the unit test for a particular
kata, if the user passes in a .spec.php file, or the whole
test suite.

We expect that most times users will want to run only the tests
for the kata they are working on.

Most of the time (and always for this base repository) we would
expect the tests to fail, and so this script is separate from
script/test, which is expected to pass in CI and elsewhere.